add BASolver class for local BA but is not good

+++ src/utils.cpp
@@ -107,7 +107,7 @@ namespace VISG {
 			R(1, 0), R(1, 1), R(1, 2), t.y(),
 			R(2, 0), R(2, 1), R(2, 2), t.z(),
 			0, 0, 0, 1;
-		return T.cast<float>();
+		return T;
 	}
 
 	void HPose2Rt(const Eigen::Matrix4f &T, Eigen::Matrix3f &R, Eigen::Vector3f &t) {
